Use createOutputFileFromCodebaseFragment(...)
Previously, `createOutputFileFromCodebase(...)` was being used by some
code which had not yet been switched to use `CodebaseFragment`. The
reason they had not yet been switched was they do not need to use the
snapshot mechanism which is built on top of it. However, that makes it
more complicated. So, this change switches them to use
`CodebaseFragment` in preparation for a follow-up change will remove
`createOutputFileFromCodebase(...)` and simplify the code.

Bug: 357833584
Test: ./gradlew
Change-Id: Ic10047b6b8f504777f7baebe8241ee07940f0924