[Refactor prep] Rename reporter to filteredReporter in APILintBase
The name makes reading the Base implementations confusing as it is not
immediately obvious where the filtering logic lies, as reporter is used
for both an object that just knows how to print (the constructor
paramter) AND an object that delegates to a reporter based on
conditional logic (the base class protected variable). Which was
initially called "reporter" but it actually takes in and delegates to a reporter that
actually does the "reporting".

Next:
Eventually this filtered reporting logic will be removed from FlaggedAPI
entirely to decouple the logic of which items to visit and what to do
with them when visited. This change prepares for that by making the code
easier to reason about while refactoring.

Bug: 442395516
Test: FlaggedApiLintTest.kt & ApiLintTest.kt
Change-Id: Iaed6bfe3a48713a7bb2c287019db7c1afe404f46