Print single line comments on a single line
Previously, single line comments like `/** @deprecated */` and
`/** Some text. */` were printed on multiple lines. That will cause
lots of differences in the formatting of Android stub files as the
current code basically copies the Javadoc straight through from the
sources and the sources typically use a single line where possible,
particularly where an automatic formatter is used.

This change attempts to replicate that behavior by only using a
multi-line comment for comments that require multiple lines. It
determines that by checking the structure of the comment and then if
necessary checking for the presence of newline characters in the
`JavadocContent`. It does not try and reformat lines as that has the
potential to corrupt the output, e.g. when it is inside a HTML `<pre>`
block.

Bug: 429965593
Test: ./gradlew
Change-Id: I03f0aa08d058f39e4c65e30a9c8eceda328a89c8