Fix handling of JvmMultifileClass split between KMP source sets
For a JvmMultifileClass, a PsiClass is generated for each source file making up the class. When the files are all in the same source set, the PsiClasses are identical, so metalava only procesed one of them.

However, if the files are in different KMP source sets and K2 is used, the PsiClasses are not identical: the PsiClass(es) from the common source set file(s) contain only the common definitions, and the PsiClass(es) from the jvm/android source set contain both common and jvm/android definitions. This meant that whether all definitions ended up in the signature file depended on whether the jvm/android or common version was processed first; if common was processed first, some definitions were missing.

This updates how JvmMultifileClasses are processed to add any additional definitions from each PsiClass as needed.

This will enable switching an AndroidX project to K2.

Bug: 417699607
Test: updated existing KMP test in UastTestBase, a test in ApiFileTest handles the non-KMP case
Change-Id: Iec18920a8128d677e16a1f8eeb8ed51b52e03e19