Generate stub documentation from the DocComment model
Previously, stub documentation was generated by printing the result of
calling `fullyQualifiedDocumentation()`. This changes it to parse the
resulting text into a `DocComment` and prints that instead.

That results in a few changes to some existing tests due to formatting
the comment more consistently and not mirroring the inconsistent source
format, e.g.
* Single line comments are printed on a single line.
* Indentation of block tags is one space not two.
* A blank link is inserted between the main description and the block
  tags if the block tags occupy more than one line.
* Trailing newlines/blank lines are removed from between block tags.

Bug: 429965593
Test: ./gradlew
Change-Id: I7a4470ece1ccdec51cefb99dff46145f4f673f6f