Add test for what should be an invalid comment
The `JavadocParser` accepts just about any content within a comment,
erroneously that includes `*/` which is the token that ends a comment
and cannot appear within a comment. This change adds a couple of tests
to show that a `*/` is currently allowed in both the main description
and in a block tag. That will be addressed in follow-up changes.

Bug: 429965593
Test: ./gradlew
Change-Id: I5be44bd1b93fd04b1564a683475933a14dcb677d