Optimize and simplify trimDocIndent(...)
Previously, `trimDocIndent(...)` would split the text into separate
lines and then iterate over the lines creating a `COMMENT_LINE`
`Matcher` for each line. This change merges that together and just
creates one `COMMENT_LINE` `Matcher` for the whole block and then
repeatedly applies it until it no longer finds a match. That has the
effect of applying it to each line in the comment.

Bug: 448617351
Test: ./gradlew
Change-Id: Ia154ddd303dd32ecceb2d41676d216711975d388