Add --Xuse-k1-uast flag to metalava
Currently, metalava defaults to using K1 unless the --Xuse-k2-uast flag
is used. We want to flip this default to K2, and start deprecating K1
support. Lint is also deprecating K1 (b/448273436).

--Xuse-k2-uast can't easily be modified to enable opting-out of K2 since
it is just a flag and isn't currently used with any arguments, so this
adds a new flag to specify K1 should be used (this mirrors the argument
being added to lint: ag/36077000).

Once this is merged, these are the next steps:
- Add the new flag for the AndroidX projects that still use K1
- Merge to main, and add the new flag for platform metalava usages
- Flip the default in metalava from K1 to K2 when neither flag is used

Bug: 445236354
Test: OptionsTest verifies that the K1 and K2 flags can't be used together
Change-Id: I922c572082079186c9e136082dd9a4ac84ac5ce5