Simplify DocAnalyzer.addDoc(...) selection of target section
Previously, it would use the `tag` parameter to search the comment for
a block tag that starts with `@$tag` and then it would assert that it
did start with that and then check to see whether the block tag started
with one of `@returnDoc`, `@paramDoc`, or `@memberDoc`. That was either
unnecessary or inefficient.

* The assertion is basically testing that `findTagDocumentation(tag)`
  works but that is actually tested elsewhere so the assertion is
  unnecessary.

* Instead of checking to see if `taggedText` starts with `@paramDoc`,
  etc. it is simpler to just check to see if `tag` matches `@paramDoc`,
  etc.

* `@memberDoc` and `@classDoc` both behave the same, i.e. `taggedText`
  is appended to the `item`'s main description so there is no need to
  handle `@memberDoc` separately.

Preparation for switching it to work on the `DocComment` model.

Bug: 450228132
Test: ./gradlew
Change-Id: I4d2216c074648baba12cbbbd4e8d24448d40bb55