Add generated version of Kotlin functions with optional parameters
When a Kotlin function has parameters with default values, the compiler generates a version which ends up used when the source function is called without all parameters from Kotlin source. Currently, metalava is not tracking the compiler generated default version, but it should be since it would be binary breaking if the signature were to change. For more details of the issue see this document: https://docs.google.com/document/d/1E1gYBNsgEUyduiEQ7yMd0o1DQ224vCNVtwqxPfBts_I/edit

Bug: 432049122
Test: CommonTargetLanguageTest, AndroidX changes: aosp/3789967
Change-Id: Ie1e01ebf864194ba89570b0cf92850aeca3ac7c5