Avoid StackOverflowError when modifying invalid documentation
Recent changes to `AbstractItemDocumentation` have changed the behavior
of the `_text` backing field. Previously, it was only `null` before it
had been initialized for the first time but those changes would also
set it to `null` when `docComment` became the source of truth for the
documentation. Unfortunately, `PsiItemDocumentation` relied on the
previous behavior which resulted in a `StackOverflowError`.

This change fixes that by adding the `AbstractItemDocumentation` method
`ensureTextBackingFieldHasBeenInitialized()` that handles the current
behavior of `_text` and abstracts the details away from
`PsiItemDocumentation`.

Bug: 450228132
Test: ./gradlew
Change-Id: I140ccc65b28dd850533c016e064e22afa0c17fa3