Add String.containsWord(String)
Used by follow-up changes. Part of the preparation work to switching
to using the `JavadocContent` model instead of treating documentation
as a `String`. Providing this method will allow a number of places
that currently operate on the documentation as a `String` using
different methods to be migrated to using the same method. That will
make it easier to migrate those uses to an equivalent method on the
`JavadocContent` model.

Bug: 450228132
Test: ./gradlew
Change-Id: I97019442ce86778fead795cf6ce96edb01f6183b