Normalize leading whitespace in descriptions
Previously, only newlines were removed from the start of descriptions.
That led to inconsistent whitespace at the beginning of the main and
block tag descriptions. This change removes all leading whitespace at
the start and then adds a space at the beginning of the main
description to match the standard Javadoc formatting.

Bug: 429965593
Test: ./gradlew
Change-Id: I430951e7ae0b389d2d5aaccd2fc5fc325572c2f7