Fully qualify link text correctly for wrapped {@link} tags
Previously, an `{@link ...}` tag without custom link text that
references a class member and is wrapped across multiple lines would
use whitespace as the custom link text when fully qualified instead
of the reference text.

e.g.
```
/**
 * {@link
 *     Other#method()}
 */
```

Would be fully qualified to:
```
/**
 * {@link test.pkg.Other#method()      }
 */
```

That was happening because it searched from the end of the content of
the `{@link}` tag and would ignore `PsiWhitespace` and
`PsiDocMethodOrFieldRef` and then return the element it found as long
as it was a `PsiDocToken`. However, when it was wrapped the leading
spaces are represented as a `PsiDocToken` not a `PsiWhitespace` so
has the following tokens:
* `PsiWhitespace` ("\n")
* `PsiDocToken` ("     ")
* `PsiDocMethodOrFieldRef`

So, it stopped at the `PsiDocTag` and thought that it was a custom link
text provided by the developer so used that instead of creating one
from the reference text.

The code should have stopped at the `PsiDocMethodOrFieldRef` as that
can never be part of custom link text so if it reaches that point then
there is no custom link text. As `PsiDocMethodOrFieldRef` is not a
`PsiDocToken` that returns `null` for the custom link text which will
cause it to generate one from the reference text.

This change fixes that by stopping ignoring `PsiDocMethodOrFieldRef`.

The original code that ignored `PsiDocMethodOrFieldRef` was added in
https://r.android.com/1550994 and there is no indication why that was
included. The test added as part of that has not been modified since
and still passes without that test.

If the link tag has custom link text at the start of a new line then
the leading spaces are kept as part of that link text. This fixes that
by trimming them from the start of the custom link text.

Bug: 429965593
Test: ./gradlew
Change-Id: Ia3315f22fec97fcf89724883a6f65b1de2976af8