Add DocDescription.content property
Previously, the `content` property of `AbstractDocDescription` was not
exposed through the `DocDescription` interface. That was an intentional
decision to minimize exposure of `JavadocContent` classes. Although it
did require down casting to `AbstractDocDescription` in tests.

Exposing the `content` property in `DocDescription` does simplify the
test and will make future changes simpler.

Bug: 450228132
Test: ./gradlew
Change-Id: I030a69ef20b2d1c4664dafc70c6ceded16501bb5