Correct performance issue around SourceFile.filterImports(...)
Change https://r.android.com/2953610 moved some functionality from
`PsiSourceFile` to `SourceFile` as part of the work to support Turbine
but unfortunately, it inadvertently introduced a performance issue.

Prior to that change there was a global cache of `Regex` objects that
was populated and reused by `docContainsWord(...)" but that change made
the cache per invocation of the method which makes it worse than
useless. This change rectifies that mistake.

Bug: 322978852
Test: ./gradlew
Change-Id: I8f7e8bb149cf3dd06394b7a7168616db0623ded8