Use containsWord(...) in AbstractSourceFile.filterImports(...)
Replaces `docContainsWord(...)` and `Regex` cache with a call to
`String.containsWord(...)`. The `docContainsWord(...)` has the overhead
of managing the `Regex` cache as well as the overhead of using the
`Regex` cache. That overhead is mitigated by checking to make sure that
the word is in the `String` in the first place but the work done to
perform that check is just wasted. The `containsWord(...)` method
avoids the use of `Regex` by searching for each occurrence of the word
and then checking to make sure it is not part of a bigger word.
As a result it has a lower overhead.

Bug: 450228132
Test: ./gradlew
Change-Id: Ic78f028fbb015f077994cacea7193069883610e3