Add TagType abstraction to support tag specific behavior
Currently, all block and inline tags behave the same way, they have a
name and some content. This change adds the `TagType` abstraction to
enable tag specific behavior, e.g. extracting the parameter name from
the content of an `@param` tag, or extracting and resolving a reference
from the content of an `@link` tag.

The `TagType` has a type parameter of type `D: TagType` that is not
currently used but is added as part of this to avoid a follow-up change
having to replace uses of `TagType` with `TagType<*>` in all the places
that this change used them on top of the other changes it makes. Doing
it here reduces the overall size of the changes.

Bug: 452699372
Test: ./gradlew
Change-Id: I5cc339f6bafb311af168bd5c22223ac2a333dc6f