Get DocContentOwner for a specific parameter
Previously, it was only possible to get `DocContentOwner` for the main
description and the first block tag of a specific type. This change
adds support for getting it from a specific `@param` block tag by
parameter name.

This test reveals a slight issue with the handling of `@param` tags as
the parameter name is not removed from the description content. That
will be fixed in a future change.

Bug: 450228132
Test: ./gradlew
Change-Id: Ia8e7d4f4b57454beb842ac84ca9ffc34aecefb3d