Reorganize LazyDocDescription parameters
Previously, it had required parameter `reporter` after optional
parameters `startInclusive` and `endExclusive` requiring the former
to be specified by name if the latter were not provided. This change
moves `reporter` first in the list to avoid that. It also dedups some
test code.

Preparation for future changes which will add extra parameters.

Bug: 450228132
Test: ./gradlew
Change-Id: Ib13b46b25514043506bdbf484a63c5e761fea2ed