Prevent metalava failure on new package added.
Metalava compatibility check will fail if a new package is added, even
if that package only has an experimental member (for more details see
the referenced bug below). This commit adds the compatibility check
functionality to see if a package contains only experimental members,
and if so won't raise an error.

Bug: 404795417
Test: `./gradlew`
Change-Id: Ieb6d4debff8c741efc2a3df7065e82c7bdfae5ec